chore: remove validator
