--- api/src/ilos/validator/index.ts
@@ -1,4 +1,0 @@
-export type { KeywordDefinition } from "@/deps.ts";
-export { AjvValidator } from "./AjvValidator.ts";
-export { ValidatorExtension } from "./ValidatorExtension.ts";
-export { ValidatorMiddleware } from "./ValidatorMiddleware.ts";
