+++ api/src/pdc/providers/validator/formats/bicCustomFormat.ts
@@ -1,5 +1,5 @@
-import { Format } from "@/ilos/validator/index.ts";
 import { isValidBIC } from "@/deps.ts";
+import { Format } from "@/pdc/providers/validator/index.ts";
 
 export const bicCustomFormat: Format = (data: string): boolean => {
   return isValidBIC(data);
