+++ api/src/pdc/providers/validator/formats/euVatCustomFormat.ts
@@ -1,4 +1,4 @@
-import { Format } from "@/ilos/validator/index.ts";
+import { Format } from "@/pdc/providers/validator/index.ts";
 
 export const euVatCustomFormat: Format = (data: string): boolean => {
   return /^[A-Z]{2}[A-Z0-9]{2}[0-9]{9}$/.test(data);
