+++ api/src/pdc/providers/validator/formats/ibanCustomFormat.ts
@@ -1,5 +1,5 @@
-import { Format } from "@/ilos/validator/index.ts";
 import { isValidIBAN } from "@/deps.ts";
+import { Format } from "@/pdc/providers/validator/index.ts";
 
 export const ibanCustomFormat: Format = (data: string): boolean => {
   return isValidIBAN(data);
