+++ api/src/pdc/providers/validator/formats/inseeCustomFormat.ts
@@ -1,4 +1,4 @@
-import { Format } from "@/ilos/validator/index.ts";
+import { Format } from "@/pdc/providers/validator/index.ts";
 
 export const inseeCustomFormat: Format = (data: string): boolean => {
   try {
