+++ api/src/pdc/providers/validator/formats/nafCustomFormat.ts
@@ -1,4 +1,4 @@
-import { Format } from "@/ilos/validator/index.ts";
+import { Format } from "@/pdc/providers/validator/index.ts";
 
 export const nafCustomFormat: Format = (data: string): boolean => {
   return /^[0-9]{4}[A-Z]{1}$/.test(data);
