+++ api/src/pdc/providers/validator/formats/nicCustomFormat.ts
@@ -1,4 +1,4 @@
-import { Format } from "@/ilos/validator/index.ts";
+import { Format } from "@/pdc/providers/validator/index.ts";
 
 export const nicCustomFormat: Format = (data: string): boolean => {
   return /^[0-9]{5}$/.test(data);
