+++ api/src/pdc/providers/validator/formats/objectidCustomFormat.ts
@@ -1,4 +1,4 @@
-import { Format } from "@/ilos/validator/index.ts";
+import { Format } from "@/pdc/providers/validator/index.ts";
 
 export const objectidCustomFormat: Format = (data: string): boolean => {
   return /^[a-f\d]{24}$/i.test(data);
