+++ api/src/pdc/providers/validator/formats/phoneCustomFormat.ts
@@ -1,4 +1,4 @@
-import { Format } from "@/ilos/validator/index.ts";
+import { Format } from "@/pdc/providers/validator/index.ts";
 import { isValidPhone } from "../lib/phone.ts";
 
 /**
