+++ api/src/pdc/providers/validator/formats/rnaCustomFormat.ts
@@ -1,4 +1,4 @@
-import { Format } from "@/ilos/validator/index.ts";
+import { Format } from "@/pdc/providers/validator/index.ts";
 
 export const rnaCustomFormat: Format = (data: string): boolean => {
   return /^W[0-9]{9}$/.test(data);
