+++ api/src/pdc/providers/validator/formats/siretCustomFormat.ts
@@ -1,4 +1,4 @@
-import { Format } from "@/ilos/validator/index.ts";
+import { Format } from "@/pdc/providers/validator/index.ts";
 
 export const siretCustomFormat: Format = (data: string): boolean => {
   return /^[0-9]{14}$/.test(data);
