+++ api/src/pdc/providers/validator/index.ts
@@ -1,8 +1,10 @@
-export { AjvValidator, ValidatorMiddleware } from "@/ilos/validator/index.ts";
-
-export type { ValidatorInterface } from "@/ilos/common/index.ts";
 export { ValidatorInterfaceResolver } from "@/ilos/common/index.ts";
+export type { ValidatorInterface } from "@/ilos/common/index.ts";
 
 export { ValidatorExtension } from "./ValidatorExtension.ts";
 
 export type * from "./types.ts";
+
+export type { Format, KeywordDefinition } from "@/deps.ts";
+export { AjvValidator } from "./AjvValidator.ts";
+export { ValidatorMiddleware } from "./ValidatorMiddleware.ts";
