+++ api/src/pdc/providers/validator/keywords/macroKeyword.ts
@@ -1,37 +1,37 @@
-import { KeywordDefinition } from "@/ilos/validator/index.ts";
+import { KeywordDefinition } from "../index.ts";
 
 import { base64Macro } from "./macros/base64Macro.ts";
-import { lonMacro } from "./macros/lonMacro.ts";
-import { latMacro } from "./macros/latMacro.ts";
 import { bicMacro } from "./macros/bicMacro.ts";
+import { countryMacro } from "./macros/countryMacro.ts";
+import { dbidMacro } from "./macros/dbidMacro.ts";
+import { departmentMacro } from "./macros/departmentMacro.ts";
+import { emailMacro } from "./macros/emailMacro.ts";
 import { euVatMacro } from "./macros/euVatMacro.ts";
 import { groupMacro } from "./macros/groupMacro.ts";
 import { ibanMacro } from "./macros/ibanMacro.ts";
 import { inseeMacro } from "./macros/inseeMacro.ts";
-import { departmentMacro } from "./macros/departmentMacro.ts";
-import { countryMacro } from "./macros/countryMacro.ts";
+import { integerMacro } from "./macros/integerMacro.ts";
+import { jwtMacro } from "./macros/jwtMacro.ts";
+import { latMacro } from "./macros/latMacro.ts";
+import { lonMacro } from "./macros/lonMacro.ts";
+import { longcharMacro } from "./macros/longcharMacro.ts";
 import { nafMacro } from "./macros/nafMacro.ts";
 import { nicMacro } from "./macros/nicMacro.ts";
-import { dbidMacro } from "./macros/dbidMacro.ts";
+import { passwordMacro } from "./macros/passwordMacro.ts";
+import { permissionsMacro } from "./macros/permissionsMacro.ts";
 import { phoneMacro } from "./macros/phoneMacro.ts";
 import { phonetruncMacro } from "./macros/phonetruncMacro.ts";
 import { postcodeMacro } from "./macros/postcodeMacro.ts";
-import { sirenMacro } from "./macros/sirenMacro.ts";
-import { siretMacro } from "./macros/siretMacro.ts";
-import { timestampMacro } from "./macros/timestampMacro.ts";
-import { varcharMacro } from "./macros/varcharMacro.ts";
-import { emailMacro } from "./macros/emailMacro.ts";
-import { passwordMacro } from "./macros/passwordMacro.ts";
 import { rnaMacro } from "./macros/rnaMacro.ts";
 import { roleMacro } from "./macros/roleMacro.ts";
 import { serialMacro } from "./macros/serialMacro.ts";
+import { sirenMacro } from "./macros/sirenMacro.ts";
+import { siretMacro } from "./macros/siretMacro.ts";
+import { timestampMacro } from "./macros/timestampMacro.ts";
 import { tokenMacro } from "./macros/tokenMacro.ts";
 import { tzMacro } from "./macros/tzMacro.ts";
 import { uuidMacro } from "./macros/uuidMacro.ts";
-import { longcharMacro } from "./macros/longcharMacro.ts";
-import { permissionsMacro } from "./macros/permissionsMacro.ts";
-import { integerMacro } from "./macros/integerMacro.ts";
-import { jwtMacro } from "./macros/jwtMacro.ts";
+import { varcharMacro } from "./macros/varcharMacro.ts";
 
 const macroStore = {
   base64: base64Macro,
